add SpanAttribute
